Changed keys to numbers + found better characters
Changed the long list to 4 hotkeys and the best ones. #1 is for after looting BT, shipwreck, or blacksmith, #2 is for bastion crafts like carrots and beds, #3 is misc stuff you can mess around with (good for BT because shears and shovel), #4 is all the extra stuff in bastions, and you can press both side mouse buttons quickly for anchors and glowstone. Gonna test tomorrow.